website/requirements.txt
flask
gunicorn
python-dotenv
yt-dlp
